+++ .storybook/theme.ts
@@ -9,6 +9,6 @@ export default create({
   brandUrl: 'https://github.com/Shopify/polaris-viz',
   brandImage: logo,
   fontBase:
-    "-apple-system, BlinkMacSystemFont, 'San Francisco', 'Segoe UI', Roboto, 'Helvetica Neue', sans-serif",
+    "Inter, -apple-system, 'system-ui', 'San Francisco', 'Segoe UI', Roboto, 'Helvetica Neue', sans-serif",
   fontCode: "Monaco, Consolas, 'Lucida Console', monospace",
 });